refactor(computer): distill the resume protocol (P3 cut 2/3)
The retry-budget loop, DNS re-registration discipline, paused-state
detection, and the paused/resume wire vocabulary move to
arcbox_computer::resume; the SandboxHost seam gains register_dns.
arcbox-api's sandbox_resume shrinks to the wire adapter (the opt-out
header and ConnectError mapping), and the engine_agent bridge retires —
call sites use SandboxHost::agent, which speaks EngineError natively.
