+++ app/arcbox-api/src/connect/filesystem.rs
@@ -19,6 +19,7 @@ use crate::ApiError;
 
 use super::sandbox_resume;
 use super::{ConnectRuntimeExt as _, ContextExt as _, with_keepalive};
+use arcbox_computer::SandboxHost as _;
 use arcbox_computer::locks::SandboxOperationLocks;
 
 /// Filesystem service implementation.
@@ -184,7 +185,7 @@ impl pb::SandboxFilesystemService for SandboxFilesystemServiceImpl {
             || {
                 let req = req.clone();
                 async {
-                    let mut agent = sandbox_resume::engine_agent(runtime, &machine)?;
+                    let mut agent = runtime.agent(&machine)?;
                     agent.sandbox_stat(req).await
                 }
             },
@@ -210,7 +211,7 @@ impl pb::SandboxFilesystemService for SandboxFilesystemServiceImpl {
             || {
                 let req = req.clone();
                 async {
-                    let mut agent = sandbox_resume::engine_agent(runtime, &machine)?;
+                    let mut agent = runtime.agent(&machine)?;
                     agent.sandbox_list_dir(req).await
                 }
             },
@@ -236,7 +237,7 @@ impl pb::SandboxFilesystemService for SandboxFilesystemServiceImpl {
             || {
                 let req = req.clone();
                 async {
-                    let mut agent = sandbox_resume::engine_agent(runtime, &machine)?;
+                    let mut agent = runtime.agent(&machine)?;
                     agent.sandbox_make_dir(req).await
                 }
             },
@@ -262,7 +263,7 @@ impl pb::SandboxFilesystemService for SandboxFilesystemServiceImpl {
             || {
                 let req = req.clone();
                 async {
-                    let mut agent = sandbox_resume::engine_agent(runtime, &machine)?;
+                    let mut agent = runtime.agent(&machine)?;
                     agent.sandbox_remove_entry(req).await
                 }
             },
@@ -288,7 +289,7 @@ impl pb::SandboxFilesystemService for SandboxFilesystemServiceImpl {
             || {
                 let req = req.clone();
                 async {
-                    let mut agent = sandbox_resume::engine_agent(runtime, &machine)?;
+                    let mut agent = runtime.agent(&machine)?;
                     agent.sandbox_move_entry(req).await
                 }
             },
